+++ backend/database/models/__init__.py
@@ -1,0 +1,15 @@
+from database.models.phong_ban import PhongBan
+from database.models.truong_dai_hoc import TruongDaiHoc
+from database.models.chuong_trinh import ChuongTrinhThucTap
+from database.models.nguoi_dung import NguoiDung
+from database.models.ho_so import HoSoThucTap
+from database.models.tai_lieu import TaiLieuHoSo
+
+__all__ = [
+    "PhongBan",
+    "TruongDaiHoc",
+    "ChuongTrinhThucTap",
+    "NguoiDung",
+    "HoSoThucTap",
+    "TaiLieuHoSo",
+]